[LLHD] Make TimeAttr track physical time as uint64_t
Since we'll want to use the LLHD `TimeAttr` to represent delays coming
from Verilog input, we'll need a fairly large range of values. Switch
the `time` field of the attribute to be a `uint64_t` to match Verilog's
underlying time type.

We'll want to rework LLHD's time type and attributes a bit in the
future, but this is a useful stepping stone.

+++ lib/Dialect/LLHD/IR/LLHDOps.cpp
@@ -55,7 +55,7 @@ OpFoldResult llhd::ConstantTimeOp::fold(FoldAdaptor adaptor) {
 }
 
 void llhd::ConstantTimeOp::build(OpBuilder &builder, OperationState &result,
-                                 unsigned time, const StringRef &timeUnit,
+                                 uint64_t time, const StringRef &timeUnit,
                                  unsigned delta, unsigned epsilon) {
   auto *ctx = builder.getContext();
   auto attr = TimeAttr::get(ctx, time, timeUnit, delta, epsilon);
